requirements.txt
tqdm
python-dotenv
beautifulsoup4